Remove dependency on metrics module from soon-to-be-stable modules
