+++ .github/dependabot.yml
@@ -188,6 +188,16 @@ updates:
       interval: weekly
   -
     package-ecosystem: gomod
+    directory: /internal/metric
+    labels:
+      - dependencies
+      - go
+      - "Skip Changelog"
+    schedule:
+      day: sunday
+      interval: weekly
+  -
+    package-ecosystem: gomod
     directory: /internal/tools
     labels:
       - dependencies
