+++ internal/metric/global/registry_test.go
