+++ internal/metric/go.mod
@@ -3,6 +3,7 @@ module go.opentelemetry.io/otel/internal/metric
 go 1.15
 
 require (
+	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/otel v0.20.0
 	go.opentelemetry.io/otel/metric v0.20.0
 )
