+++ metric/oteltest/meter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package oteltest // import "go.opentelemetry.io/otel/oteltest"
+package oteltest // import "go.opentelemetry.io/otel/metric/oteltest"
 
 import (
 	"context"
