+++ oteltest/go.mod
@@ -51,8 +51,6 @@ replace go.opentelemetry.io/otel/trace => ../trace
 require (
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/otel v0.20.0
-	go.opentelemetry.io/otel/internal/metric v0.20.0
-	go.opentelemetry.io/otel/metric v0.20.0
 	go.opentelemetry.io/otel/trace v0.20.0
 )
